changed from int to florat for reading the high score file, added esc to quit

+++ game_stats.py
@@ -13,7 +13,7 @@ class GameStats:
         # Try to see if there's a high score record.
         try:
             with open("high_score.txt") as file_obj:
-                self.high_score = int(file_obj.read())
+                self.high_score = float(file_obj.read())
         except FileNotFoundError:
             self.high_score = 0
